Change name to "zscope"
"zless" is already a utility installed on many Unix systems; I'm not
sure if I purposely co-opted that name or it was an accident, but to
avoid name clashes, I am changing the name of the project to "zscope".

To avoid any references to the old name, I regenerated all test
fixtures, which required updating some tests to accommodate the name
change as well.